Move various sdist functionality back out of flit_core
