+++ flit_core/flit_core/__init__.py
@@ -1,0 +1,7 @@
+"""Flit's core machinery for building packages.
+
+This package provides a standard PEP 517 API to build packages using Flit.
+All the convenient development features live in the main 'flit' package.
+"""
+
+__version__ = '0.1'
