+++ flit_core/flit_core/sdist.py
@@ -6,31 +6,12 @@ import logging
 import os
 from pathlib import Path
 from posixpath import join as pjoin
-from pprint import pformat
 import tarfile
 
 from . import common, inifile
-from .common import VCSError
-from flit.vcs import identify_vcs
 
 log = logging.getLogger(__name__)
 
-SETUP = """\
-#!/usr/bin/env python
-# setup.py generated by flit for tools that don't yet use PEP 517
-
-from distutils.core import setup
-
-{before}
-setup(name={name!r},
-      version={version!r},
-      description={description!r},
-      author={author!r},
-      author_email={author_email!r},
-      url={url!r},
-      {extra}
-     )
-"""
 
 PKG_INFO = """\
 Metadata-Version: 1.1
@@ -42,97 +23,6 @@ Author: {author}
 Author-email: {author_email}
 """
 
-
-def auto_packages(pkgdir: str):
-    """Discover subpackages and package_data"""
-    pkgdir = os.path.normpath(pkgdir)
-    pkg_name = os.path.basename(pkgdir)
-    pkg_data = defaultdict(list)
-    # Undocumented distutils feature: the empty string matches all package names
-    pkg_data[''].append('*')
-    packages = [pkg_name]
-    subpkg_paths = set()
-
-    def find_nearest_pkg(rel_path):
-        parts = rel_path.split(os.sep)
-        for i in reversed(range(1, len(parts))):
-            ancestor = '/'.join(parts[:i])
-            if ancestor in subpkg_paths:
-                pkg = '.'.join([pkg_name] + parts[:i])
-                return pkg, '/'.join(parts[i:])
-
-        # Relative to the top-level package
-        return pkg_name, rel_path
-
-    for path, dirnames, filenames in os.walk(pkgdir, topdown=True):
-        if os.path.basename(path) == '__pycache__':
-            continue
-
-        from_top_level = os.path.relpath(path, pkgdir)
-        if from_top_level == '.':
-            continue
-
-        is_subpkg = '__init__.py' in filenames
-        if is_subpkg:
-            subpkg_paths.add(from_top_level)
-            parts = from_top_level.split(os.sep)
-            packages.append('.'.join([pkg_name] + parts))
-        else:
-            pkg, from_nearest_pkg = find_nearest_pkg(from_top_level)
-            pkg_data[pkg].append(pjoin(from_nearest_pkg, '*'))
-
-    # Sort values in pkg_data
-    pkg_data = {k: sorted(v) for (k, v) in pkg_data.items()}
-
-    return sorted(packages), pkg_data
-
-def _parse_req(requires_dist):
-    """Parse "Foo (v); python_version == '2.x'" from Requires-Dist
-
-    Returns pip-style appropriate for requirements.txt.
-    """
-    if ';' in requires_dist:
-        name_version, env_mark = requires_dist.split(';', 1)
-        env_mark = env_mark.strip()
-    else:
-        name_version, env_mark = requires_dist, None
-
-    if '(' in name_version:
-        # turn 'name (X)' and 'name (<X.Y)'
-        # into 'name == X' and 'name < X.Y'
-        name, version = name_version.split('(', 1)
-        name = name.strip()
-        version = version.replace(')', '').strip()
-        if not any(c in version for c in '=<>'):
-            version = '==' + version
-        name_version = name + version
-
-    return name_version, env_mark
-
-def convert_requires(reqs_by_extra):
-    """Regroup requirements by (extra, env_mark)"""
-    grouping = defaultdict(list)
-    for extra, reqs in reqs_by_extra.items():
-        for req in reqs:
-            name_version, env_mark = _parse_req(req)
-            grouping[(extra, env_mark)].append(name_version)
-
-    install_reqs = grouping.pop(('.none',  None), [])
-    extra_reqs = {}
-    for (extra, env_mark), reqs in grouping.items():
-        if extra == '.none':
-            extra = ''
-        if env_mark is None:
-            extra_reqs[extra] = reqs
-        else:
-            extra_reqs[extra + ':' + env_mark] = reqs
-
-    return install_reqs, extra_reqs
-
-def include_path(p):
-    return not (p.startswith('dist' + os.sep)
-                or (os.sep+'__pycache__' in p)
-                or p.endswith('.pyc'))
 
 def clean_tarinfo(ti, mtime=None):
     """Clean metadata from a TarInfo object to make it more reproducible.
@@ -170,55 +60,22 @@ class SdistBuilder:
 
         return dict(res)
 
-    def find_tracked_files(self):
-        vcs_mod = identify_vcs(self.srcdir)
-        untracked_deleted = vcs_mod.list_untracked_deleted_files(self.srcdir)
-        if list(filter(include_path, untracked_deleted)):
-            raise VCSError("Untracked or deleted files in the source directory. "
-                           "Commit, undo or ignore these files in your VCS.",
-                           self.srcdir)
+    def select_files(self):
+        """Pick which files from the source tree will be included in the sdist
 
-        files = vcs_mod.list_tracked_files(self.srcdir)
-        log.info("Found %d files tracked in %s", len(files), vcs_mod.name)
-        return sorted(filter(include_path, files))
+        This is overridden in flit itself to use information from a VCS to
+        include tests, docs, etc. for a 'gold standard' sdist.
+        """
+        return list(self.module.iter_files()) + [self.ini_path.name] \
+              + self.ini_info['referenced_files']
 
-    def make_setup_py(self):
-        before, extra = [], []
-        if self.module.is_package:
-            packages, package_data = auto_packages(str(self.module.path))
-            before.append("packages = \\\n%s\n" % pformat(sorted(packages)))
-            before.append("package_data = \\\n%s\n" % pformat(package_data))
-            extra.append("packages=packages,")
-            extra.append("package_data=package_data,")
-        else:
-            extra.append("py_modules={!r},".format([self.module.name]))
+    def add_setup_py(self, files_to_add, target_tarfile):
+        """No-op here; overridden in flit to generate setup.py"""
+        pass
 
-        install_reqs, extra_reqs = convert_requires(self.ini_info['reqs_by_extra'])
-        if install_reqs:
-            before.append("install_requires = \\\n%s\n" % pformat(install_reqs))
-            extra.append("install_requires=install_requires,")
-        if extra_reqs:
-            before.append("extras_require = \\\n%s\n" % pformat(extra_reqs))
-            extra.append("extras_require=extras_require,")
-
-        entrypoints = self.prep_entry_points()
-        if entrypoints:
-            before.append("entry_points = \\\n%s\n" % pformat(entrypoints))
-            extra.append("entry_points=entry_points,")
-
-        if self.metadata.requires_python:
-            extra.append('python_requires=%r,' % self.metadata.requires_python)
-
-        return SETUP.format(
-            before='\n'.join(before),
-            name=self.metadata.name,
-            version=self.metadata.version,
-            description=self.metadata.summary,
-            author=self.metadata.author,
-            author_email=self.metadata.author_email,
-            url=self.metadata.home_page,
-            extra='\n      '.join(extra),
-        ).encode('utf-8')
+    @property
+    def dir_name(self):
+        return '{}-{}'.format(self.metadata.name, self.metadata.version)
 
     def build(self, target_dir:Path =None):
         if target_dir is None:
@@ -234,13 +91,11 @@ class SdistBuilder:
                              format=tarfile.PAX_FORMAT)
 
         try:
-            tf_dir = '{}-{}'.format(self.metadata.name, self.metadata.version)
-
-            files_to_add = self.find_tracked_files()
+            files_to_add = self.select_files()
 
             for relpath in files_to_add:
                 path = self.srcdir / relpath
-                ti = tf.gettarinfo(str(path), arcname=pjoin(tf_dir, relpath))
+                ti = tf.gettarinfo(str(path), arcname=pjoin(self.dir_name, relpath))
                 ti = clean_tarinfo(ti, mtime)
 
                 if ti.isreg():
@@ -249,14 +104,7 @@ class SdistBuilder:
                 else:
                     tf.addfile(ti)  # Symlinks & ?
 
-            if 'setup.py' in files_to_add:
-                log.warning("Using setup.py from repository, not generating setup.py")
-            else:
-                setup_py = self.make_setup_py()
-                log.info("Writing generated setup.py")
-                ti = tarfile.TarInfo(pjoin(tf_dir, 'setup.py'))
-                ti.size = len(setup_py)
-                tf.addfile(ti, io.BytesIO(setup_py))
+            self.add_setup_py(files_to_add, tf)
 
             pkg_info = PKG_INFO.format(
                 name=self.metadata.name,
@@ -266,7 +114,7 @@ class SdistBuilder:
                 author=self.metadata.author,
                 author_email=self.metadata.author_email,
             ).encode('utf-8')
-            ti = tarfile.TarInfo(pjoin(tf_dir, 'PKG-INFO'))
+            ti = tarfile.TarInfo(pjoin(self.dir_name, 'PKG-INFO'))
             ti.size = len(pkg_info)
             tf.addfile(ti, io.BytesIO(pkg_info))
